数据

+++ app/Http/Controllers/ManagerController.php
@@ -15,8 +15,14 @@ class ManagerController extends Controller
      */
     public function index(Request $request)
     {
-        $data=DB::table('manager')->orderBy('id','desc')->paginate(10);
-        return view('admin.manager.index',['data'=>$data,]);
+        if (session('roles') == 1) {
+            $data=DB::table('manager')->orderBy('id','desc')->paginate(10);
+            return view('admin.manager.index',['data'=>$data,]);
+        }else if(session('roles') == 2){
+            $da = DB::table('manager')->where('name',session('admin_name'))->first();
+            return view('admin.manager.index',['da'=>$da]);
+        }
+                    
     }
 
     /**
@@ -26,7 +32,12 @@ class ManagerController extends Controller
      */
     public function create()
     {
-        return view('admin.manager.create');
+        if(session('roles') == 1){
+           return view('admin.manager.create'); 
+        }else{
+           return back()->with(['msg'=>'您不是超级管理员或权限不足,请联系超级管理员!','msg_info'=>'alert-danger']); 
+        }
+        
     }
 
     /**
@@ -39,7 +50,7 @@ class ManagerController extends Controller
     {       
        $data=$request->except(['_token','repwd']);
        $data['pwd']=Hash::make($data['pwd']);
-       $res=DB::table('manager')->where('name',$data['name'])->first();       
+       $res=DB::table('manager')->where('name',$data['name'])->first();
        if (empty($res)){
            DB::table('manager')->insert($data);
            return redirect('/admin/manager')->with(['msg'=>'ok~ 添加成功!','msg_info'=>'alert-success']);
@@ -67,8 +78,14 @@ class ManagerController extends Controller
      */
     public function edit(Request $request,$id)
     {
-        $res=DB::table('manager')->where('id',$id)->first();
-        return view('admin.manager.edit',['res'=>$res]);
+        $re = DB::table('manager')->where('id',$id)->first();        
+        if (session('admin_name') == $re->name || session('roles') == 1) {
+            $res=DB::table('manager')->where('id',$id)->first();
+            return view('admin.manager.edit',['res'=>$res]);
+        }else{
+             return back()->with(['msg'=>'您当前账号不匹配或您不是超级管理员!请联系超级管理员!','msg_info'=>'alert-danger']);
+        }
+        
     }
 
     /**
@@ -104,11 +121,17 @@ class ManagerController extends Controller
      */
     public function destroy($id)
     {
-       $res = DB::table('manager')->where('id',$id)->delete();
-       if ($res) {
-           return redirect('/admin/manager')->with(['msg'=>'ok~ 删除成功!','msg_info'=>'alert-success']);
+        $re = DB::table('manager')->where('id',$id)->first();
+       if(session('roles') == 1 && $re->roles != 1){
+            $res = DB::table('manager')->where('id',$id)->delete();
+            if ($res) {
+               return redirect('/admin/manager')->with(['msg'=>'ok~ 删除成功!','msg_info'=>'alert-success']);
+            }else{
+               return back()->with(['msg'=>'oh! 删除失败!','msg_info'=>'alert-danger']);
+            }
        }else{
-           return back()->with(['msg'=>'oh! 删除失败!','msg_info'=>'alert-danger']);
+            return back()->with(['msg'=>'您不是超级管理员或权限不足或删除管理员!请联系超级管理员!','msg_info'=>'alert-danger']);
        }
+       
     }
 }
